+++ CHANGELOG.md
@@ -367,6 +367,8 @@ This update makes several changes to Predator that dramatically improve its proc
 - Improved debug message handling when multiple threads are running concurrently.
 - Added a configuration value to disable the logging of all license plate guesses alongside the top guess.
 - Fixed some broken references to configuration values.
-- Both errors and warnings are now logged to the error log file in the interface directory, instead of just errors.
+- Updated the way errors and warnings are saved in the interface directory.
+    - Both errors and warnings are now logged to the error log file in the interface directory, instead of just errors.
 - Fixed an issue where information displayed in management mode would not wait for user input before clearing the screen.
 - Prompts that ask the user to press enter to coninue are now displayed in a faint font.
+- Made the 'plates detected' count display in real-time mode more consistent.